% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{atm_mcmc}
\alias{atm_mcmc}
\title{ATM master function}
\usage{
atm_mcmc(atm_df, n_mcmc, n_burn)
}
\arguments{
\item{atm_df}{object returned by convert_atm()}

\item{n_mcmc}{number of mcmc iterations}

\item{n_burn}{number of burn-in iterations}
}
\description{
Master function implementing an MCMC algorithm for fitting the ATM model with single kappa parameter.
}
