% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{atm_vM_1kappa}
\alias{atm_vM_1kappa}
\title{ATM model MCMC sampler}
\usage{
atm_vM_1kappa(z.list, theta.list, grid.list, theta.tmp, I_n, n.mcmc, n.burn,
  cat_n)
}
\arguments{
\item{z.list}{list of observer coordinates}

\item{theta.list}{list of observed bearings}

\item{grid.list}{list of direct approximation grid points}

\item{theta.tmp}{list of theta values corresponding to grid points}

\item{I_n}{number of locations}

\item{n.mcmc}{number of total mcmc iterations}

\item{n.burn}{number of burn-in iterations}

\item{cat_n}{mcmc iteration printing interval}
}
\description{
MCMC sampler for ATM model with single kappa parameter.
}
