% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bearing_int_avg}
\alias{bearing_int_avg}
\title{Compute average intersection}
\usage{
bearing_int_avg(obs_pts, theta)
}
\arguments{
\item{obs_pts}{observer locations}

\item{theta}{bearing taken from each observer location}
}
\description{
Function that computes average intersection of bearings
}
