% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_atm}
\alias{convert_atm}
\title{Generate ATM dataframe}
\usage{
convert_atm(df)
}
\arguments{
\item{df}{user supplied dataframe}
}
\description{
Generate objects in format usable by razimuth package functions.
}
