% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_signal.R
\name{loc_signal}
\alias{loc_signal}
\title{Locate transmitter signal - Lenth's method}
\usage{
loc_signal(df, max_iter = 100, param_tol = 0.001, robust = FALSE,
  psi_fun = c("Huber", "Andrews"), psi_c = 1.5, buffer = 100)
}
\arguments{
\item{df}{user supplied dataframe}

\item{max_iter}{maximum number of iterations}

\item{param_tol}{tolerance after which algorithm terminates iterations}

\item{robust}{logical for M-estimation}

\item{psi_fun}{psi function}

\item{psi_c}{psi function tuning constant}

\item{buffer}{buffer distance (m) extending bbox secondary location filter}
}
\description{
Estimation of transmitter location using azimuthal telemetry data and the algorithms outlined in <http://www.jstor.org/stable/1268030>
}
