% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{locate_est}
\alias{locate_est}
\title{Modified locate()}
\usage{
locate_est(x, int_avg, int_mat)
}
\arguments{
\item{x}{user supplied dataframe}

\item{int_avg}{average intersection}

\item{int_mat}{matrix containing all unique intersection points}
}
\description{
Modified version of locate() from the R package sigloc.
}
