% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{p_isopleth}
\alias{p_isopleth}
\title{Plot transmitter location posterior credible isopleth}
\usage{
p_isopleth(df, prob_lvls, range_extend, kde_n, col_vec)
}
\arguments{
\item{df}{matrix containing posterior draws}

\item{prob_lvls}{credible isopleth probability value(s)}

\item{range_extend}{buffer distance to add to range of df to prevent isopleth discontinuity}

\item{kde_n}{number of grid points in each direction - scalar n argument from kde2d()}

\item{col_vec}{vector of length prob_lvls}
}
\description{
Plotting function to plot posterior credible isopleth(s) of transmitter location.
}
