% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_kappa}
\alias{plot_kappa}
\title{Diagnostic plots for kappa parameter}
\usage{
plot_kappa(atm_obj, item = c("traceplot", "density", "tuning", "run_mean"))
}
\arguments{
\item{atm_obj}{object returned by atm_mcmc()}

\item{item}{specify which diagnostic to plot}
}
\description{
Plotting function to produce diagnostic plots of the kappa parameter MCMC output.
}
