% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.R
\name{sim_atd}
\alias{sim_atd}
\title{Simulate azimuthal telemetry data}
\usage{
sim_atd(n_loc = 100, sq_km = 4, n_azimuth = 3, dist_vec = c(200, 500),
  kappa_sim = 50, prior_r = 1000, plot = FALSE)
}
\arguments{
\item{n_loc}{number of transmitter locations to simulate uniformly in square of area sq_km}

\item{sq_km}{area of square spatial domain in km}

\item{n_azimuth}{number of azimuths taken to triangulate transmitter location}

\item{dist_vec}{min and max distance between observer and transmitter location (drawn uniformly)}

\item{kappa_sim}{value for kappa governing uncertainty in which we observe azimuths}

\item{prior_r}{prior radius}

\item{plot}{logical on whether or not to plot each simulated transmitter location with observed azimuths}
}
\description{
Simulate azimuthal telemetry data under random design for demonstration purposes
}
