% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize_atm}
\alias{visualize_atm}
\title{Visualize azimuthal telemetry data}
\usage{
visualize_atm(atm_df, obs_id, add_prior = FALSE, add_legend = TRUE)
}
\arguments{
\item{atm_df}{object returned by convert_atm()}

\item{obs_id}{observation id}

\item{add_prior}{logical specifying whether to plot circular prior}

\item{add_legend}{logical specifying whether to add legend to plot}
}
\description{
Plotting function to visualize user supplied azimuthal telemetry data.
}
