% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{which_pid}
\alias{which_pid}
\title{pid identifier helper}
\usage{
which_pid(atm_df, obs_id)
}
\arguments{
\item{atm_df}{object returned by convert_atm()}

\item{obs_id}{observation id}
}
\description{
Function to identify pid for a given obs_id
}
